Conflicts:
	Armlet_fw/SharedAtlLogic/emotions.h
	Armlet_fw/src/application.cpp

+++ MusicProcessor/EmotionProcessor.py
@@ -64,7 +64,7 @@ CSV_TARGET = 'Reasons.csv'
 
 C_CONTENT = '''\
 /*
- * emotions.c
+ * emotions.cpp
  *
  * Part of "Ticket to Atlantis" LARP music engine.
  *
@@ -180,7 +180,7 @@ extern Emotion_t emotions[];
 typedef struct Reason {
     const char* name;
     int weight;
-    int age;
+    long age;
     int eID;
 } Reason_t;
